Reads data off ttbar.root and writes to .txt
This reads off data from the events in ttbar.root. It writes to ttbar_event_output.txt the number of particles in an event, and then on one line per particle, writes px py pz E. Note that it reads off particle data from ttbar.root in the form of pt, eta, phi, m, so it has to do some conversions first.