Histogrammed deviances of jet parameters
After using the ConstituentSubtractor and IteratedConstituentSubtractor on about 10,000 events in ttbar.root, the deviance of jet parameters between the subtracted jets and the signal jets (in ttbar.root) are histogrammed. I have included:
pt
rapidity
phi
energy

Also, I included more histograms for how many additional jets the subtraction algorithms found (above the 20 GeV cutoff) that are not present in the signal jets.